pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plex-media-tool"
version = "1.0.0"
description = "Automated Plex media file renaming and transcoding pipeline with TMDb metadata lookup"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Bri Stevenski", email = "bstevenski@gmail.com" }
]
license = { text = "MIT" }
keywords = ["plex", "media", "transcoding", "tmdb", "ffmpeg", "video"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Video :: Conversion",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
]

dependencies = [
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "ffmpeg-python>=0.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
plex-rename = "rename_media_files:main"
plex-transcode = "transcode_media_files:main"

[tool.hatch.build.targets.wheel]
packages = ["src/common", "src/rename_utils", "src/transcode_utils"]

[tool.hatch.build.targets.sdist]
include = ["src/"]

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "W", "I", "N", "UP"]
lint.ignore = ["E501"]  # Line too long (handled by black)

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.hatch.envs.default.scripts]
# Media processing commands
rename = "plex-rename"
transcode = "plex-transcode"
rename-debug = "plex-rename --dry-run --log-level DEBUG"
transcode-debug = "plex-transcode --dry-run --log-level DEBUG"

# Process management commands
ps = """
     pgrep -fl 'rename_media_files|transcode_media_files' || echo 'No media tool processes found';
     pgrep -fl ffmpeg || echo 'No ffmpeg processes found'
     """
trail-log = """
            ls -t .logs/ 2>/dev/null |
            head -1 | xargs tail -f 2>/dev/null ||
            echo 'No log files found'
            """
kill = """
       pkill -f 'rename_media_files|transcode_media_files';
       pkill -f ffmpeg;
       echo 'Killed media tool and ffmpeg processes'
       """
clean-python = """
               rm -rf `find . -type d -name __pycache__` 2>/dev/null || true;
               rm -f `find . -name '*.pyc'` 2>/dev/null || true;
               rm -f `find . -name '*.pyo'` 2>/dev/null || true;
               rm -rf `find . -name '*.egg-info'` 2>/dev/null || true;
               rm -rf `find . -name '.pytest_cache'` 2>/dev/null || true;
               echo '✅ Python clean complete'
               """
clean-logs = """
             rm -rf .logs 2>/dev/null;
             echo '✅ Log clean complete'
             """
clean-all = ['clean-python', 'clean-logs']
help = """
        echo 'Available commands:'; echo
        echo '   hatch run rename'; echo '      - Run media renamer in standard mode'; echo
        echo '   hatch run transcode'; echo '      - Run media transcoder in standard mode'; echo
        echo '   hatch run rename-debug'; echo '      - Run media renamer in debug/dry-run mode'; echo
        echo '   hatch run transcode-debug'; echo '      - Run media transcoder in debug/dry-run mode'; echo
        echo '   Console scripts (after pip install -e .):'; echo
        echo '     plex-rename'; echo '        - Run media renamer directly'; echo
        echo '     plex-transcode'; echo '     - Run media transcoder directly'; echo
        echo '   hatch run ps'; echo '      - List running media tool and ffmpeg processes'; echo
        echo '   hatch run trail-log'; echo '      - Tail of latest log file'; echo
        echo '   hatch run kill'; echo '      - Stop any running media tools and ffmpeg processes'; echo
        echo '   hatch run clean-python'; echo '      - Remove Python cache files and directories'; echo
        echo '   hatch run clean-logs'; echo '      - Remove old log files'; echo
        echo '   hatch run clean-all'; echo '      - Perform a full clean (Python cache + log files)'; echo
        """

[tool.pydocstyle]
select = ["D"]
match-dir = "src"
